Add Migration

+++ First API/First API/Services/Handlers/TimeSpanHandler.cs
@@ -7,7 +7,7 @@ namespace First_API.Handlers
     public class TimeSpanHandler:SqlMapper.TypeHandler<TimeSpan>
     {
         public override TimeSpan Parse(object value)
-            => TimeSpan.FromSeconds((int)value);
+            => TimeSpan.FromSeconds((long)value);
 
         public override void SetValue(IDbDataParameter parameter, 
                                       TimeSpan value)
